refactor: split docs graph entrypoint helpers
